The function ppl_set_deterministic_timeout() in the C interface now takes an unsigned long long parameter. It erroneously used to take an unsigned int.

+++ interfaces/C/ppl_c_implementation_common.cc
@@ -259,7 +259,7 @@ ppl_reset_timeout(void) try {
 CATCH_ALL
 
 int
-ppl_set_deterministic_timeout(unsigned weight) try {
+ppl_set_deterministic_timeout(unsigned long long weight) try {
   // In case a deterministic timeout was already set.
   reset_deterministic_timeout();
   static timeout_exception e;
